Add buttons to pause/play, step through, and reset sorting
The number of array accesses and comparisons have temporarily been hidden, because I have not found a good place for them yet.

The canvas is redrawn even if no sorting algorithm is running. Although this wastes cycles, it is much easier to work with. For example, it automatically resizes if any of the parameters are changed.